Embeddings: add 'dry run' for embedding jobs
