+++ enterprise/cmd/frontend/internal/embeddings/resolvers/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "//enterprise/internal/embeddings",
         "//enterprise/internal/embeddings/background/contextdetection",
         "//enterprise/internal/embeddings/background/repo",
+        "//enterprise/internal/embeddings/embed",
         "//internal/api",
         "//internal/auth",
         "//internal/cody",
