+++ enterprise/cmd/worker/internal/embeddings/repo/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "repo",
     srcs = [
-        "document_ranks.go",
         "handler.go",
         "janitor.go",
         "worker.go",
@@ -17,16 +16,11 @@ go_library(
         "//enterprise/internal/embeddings",
         "//enterprise/internal/embeddings/background/repo",
         "//enterprise/internal/embeddings/embed",
-        "//enterprise/internal/embeddings/split",
         "//internal/actor",
-        "//internal/api",
-        "//internal/api/internalapi",
-        "//internal/codeintel/types",
         "//internal/conf",
         "//internal/env",
         "//internal/gitserver",
         "//internal/goroutine",
-        "//internal/httpcli",
         "//internal/observation",
         "//internal/uploadstore",
         "//internal/workerutil",
