+++ enterprise/internal/embeddings/embed/BUILD.bazel
@@ -4,7 +4,9 @@ go_library(
     name = "embed",
     srcs = [
         "api.go",
+        "document_ranks.go",
         "embed.go",
+        "embed_files.go",
         "files.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/embeddings/embed",
@@ -14,9 +16,11 @@ go_library(
         "//enterprise/internal/embeddings/split",
         "//enterprise/internal/paths",
         "//internal/api",
+        "//internal/api/internalapi",
         "//internal/binary",
         "//internal/codeintel/types",
         "//internal/conf",
+        "//internal/gitserver",
         "//internal/httpcli",
         "//lib/errors",
         "//schema",
