+++ enterprise/internal/embeddings/embed/document_ranks.go
@@ -1,4 +1,4 @@
-package repo
+package embed
 
 import (
 	"context"
