+++ enterprise/internal/embeddings/embed/embed.go
@@ -4,21 +4,36 @@ import (
 	"context"
 	"time"
 
-	"github.com/sourcegraph/sourcegraph/internal/codeintel/types"
-	"github.com/sourcegraph/sourcegraph/lib/errors"
-
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/embeddings"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/embeddings/split"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/paths"
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/types"
+	"github.com/sourcegraph/sourcegraph/internal/conf"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 )
 
 const GET_EMBEDDINGS_MAX_RETRIES = 5
-
-const EMBEDDING_BATCHES = 5
 const EMBEDDING_BATCH_SIZE = 512
 
 const maxFileSize = 1000000 // 1MB
+
+// The threshold to embed the entire file is slightly larger than the chunk threshold to
+// avoid splitting small files unnecessarily.
+const (
+	embedEntireFileTokensThreshold          = 384
+	embeddingChunkTokensThreshold           = 256
+	embeddingChunkEarlySplitTokensThreshold = embeddingChunkTokensThreshold - 32
+
+	defaultMaxCodeEmbeddingsPerRepo = 3_072_000
+	defaultMaxTextEmbeddingsPerRepo = 512_000
+)
+
+var splitOptions = split.SplitOptions{
+	NoSplitTokensThreshold:         embedEntireFileTokensThreshold,
+	ChunkTokensThreshold:           embeddingChunkTokensThreshold,
+	ChunkEarlySplitTokensThreshold: embeddingChunkEarlySplitTokensThreshold,
+}
 
 type ranksGetter func(ctx context.Context, repoName string) (types.RepoPathRanks, error)
 
@@ -27,10 +42,41 @@ type ranksGetter func(ctx context.Context, repoName string) (types.RepoPathRanks
 // It returns a RepoEmbeddingIndex containing the embeddings and metadata.
 func EmbedRepo(
 	ctx context.Context,
+	repoName api.RepoName,
+	revision api.CommitID,
+	client EmbeddingsClient,
+	gitserverClient gitserver.Client,
+) (*embeddings.RepoEmbeddingIndex, *embeddings.EmbedRepoStats, error) {
+	fetcher := &revisionFetcher{
+		repo:      repoName,
+		revision:  revision,
+		gitserver: gitserverClient,
+	}
+
+	config := conf.Get().Embeddings
+	excludedGlobPatterns := GetDefaultExcludedFilePathPatterns()
+	excludedGlobPatterns = append(excludedGlobPatterns, CompileGlobPatterns(config.ExcludedFilePathPatterns)...)
+
+	opts := embedRepoOptions{
+		RepoName:          repoName,
+		Revision:          revision,
+		ExcludePatterns:   excludedGlobPatterns,
+		SplitOptions:      splitOptions,
+		MaxCodeEmbeddings: defaultTo(config.MaxCodeEmbeddingsPerRepo, defaultMaxCodeEmbeddingsPerRepo),
+		MaxTextEmbeddings: defaultTo(config.MaxTextEmbeddingsPerRepo, defaultMaxTextEmbeddingsPerRepo),
+	}
+
+	return embedRepo(ctx, client, fetcher, getDocumentRanks, opts)
+}
+
+// embedRepo is a private helper method that accepts helper objects and options to control the
+// embeddings process. It is useful for unit testing.
+func embedRepo(
+	ctx context.Context,
 	client EmbeddingsClient,
 	readLister FileReadLister,
 	getDocumentRanks ranksGetter,
-	opts EmbedRepoOpts,
+	opts embedRepoOptions,
 ) (*embeddings.RepoEmbeddingIndex, *embeddings.EmbedRepoStats, error) {
 	start := time.Now()
 
@@ -81,7 +127,7 @@ func EmbedRepo(
 	return index, stats, nil
 }
 
-type EmbedRepoOpts struct {
+type embedRepoOptions struct {
 	RepoName          api.RepoName
 	Revision          api.CommitID
 	ExcludePatterns   []*paths.GlobPattern
@@ -90,130 +136,39 @@ type EmbedRepoOpts struct {
 	MaxTextEmbeddings int
 }
 
-// embedFiles embeds file contents from the given file names. Since embedding models can only handle a certain amount of text (tokens) we cannot embed
-// entire files. So we split the file contents into chunks and get embeddings for the chunks in batches. Functions returns an EmbeddingIndex containing
-// the embeddings and metadata about the chunks the embeddings correspond to.
-func embedFiles(
-	ctx context.Context,
-	files []FileEntry,
-	client EmbeddingsClient,
-	excludePatterns []*paths.GlobPattern,
-	splitOptions split.SplitOptions,
-	reader FileReader,
-	maxEmbeddingVectors int,
-	repoPathRanks types.RepoPathRanks,
-) (embeddings.EmbeddingIndex, embeddings.EmbedFilesStats, error) {
-	start := time.Now()
+func defaultTo(input, def int) int {
+	if input == 0 {
+		return def
+	}
+	return input
+}
 
-	dimensions, err := client.GetDimensions()
+type revisionFetcher struct {
+	repo      api.RepoName
+	revision  api.CommitID
+	gitserver gitserver.Client
+}
+
+func (r *revisionFetcher) Read(ctx context.Context, fileName string) ([]byte, error) {
+	return r.gitserver.ReadFile(ctx, nil, r.repo, r.revision, fileName)
+}
+
+func (r *revisionFetcher) List(ctx context.Context) ([]FileEntry, error) {
+	fileInfos, err := r.gitserver.ReadDir(ctx, nil, r.repo, r.revision, "", true)
 	if err != nil {
-		return embeddings.EmbeddingIndex{}, embeddings.EmbedFilesStats{}, err
+		return nil, err
 	}
 
-	index := embeddings.EmbeddingIndex{
-		Embeddings:      make([]int8, 0, len(files)*dimensions/2),
-		RowMetadata:     make([]embeddings.RepoEmbeddingRowMetadata, 0, len(files)/2),
-		ColumnDimension: dimensions,
-		Ranks:           make([]float32, 0, len(files)/2),
+	entries := make([]FileEntry, 0, len(fileInfos))
+	for _, fileInfo := range fileInfos {
+		if !fileInfo.IsDir() {
+			entries = append(entries, FileEntry{
+				Name: fileInfo.Name(),
+				Size: fileInfo.Size(),
+			})
+		}
 	}
-
-	var batch []split.EmbeddableChunk
-
-	flush := func() error {
-		if len(batch) == 0 {
-			return nil
-		}
-
-		batchChunks := make([]string, len(batch))
-		for idx, chunk := range batch {
-			batchChunks[idx] = chunk.Content
-			index.RowMetadata = append(index.RowMetadata, embeddings.RepoEmbeddingRowMetadata{FileName: chunk.FileName, StartLine: chunk.StartLine, EndLine: chunk.EndLine})
-
-			// Unknown documents have rank 0. Zoekt is a bit smarter about this, assigning 0
-			// to "unimportant" files and the average for unknown files. We should probably
-			// add this here, too.
-			index.Ranks = append(index.Ranks, float32(repoPathRanks.Paths[chunk.FileName]))
-		}
-
-		batchEmbeddings, err := client.GetEmbeddingsWithRetries(ctx, batchChunks, GET_EMBEDDINGS_MAX_RETRIES)
-		if err != nil {
-			return errors.Wrap(err, "error while getting embeddings")
-		}
-		index.Embeddings = append(index.Embeddings, embeddings.Quantize(batchEmbeddings)...)
-
-		batch = batch[:0] // reset batch
-		return nil
-	}
-
-	addToBatch := func(chunk split.EmbeddableChunk) error {
-		batch = append(batch, chunk)
-		if len(batch) >= EMBEDDING_BATCH_SIZE {
-			// Flush if we've hit batch size
-			return flush()
-		}
-		return nil
-	}
-
-	var (
-		statsEmbeddedByteCount  int
-		statsEmbeddedFileCount  int
-		statsEmbeddedChunkCount int
-		statsSkipped            SkipStats
-	)
-	for _, file := range files {
-		// This is a fail-safe measure to prevent producing an extremely large index for large repositories.
-		if statsEmbeddedChunkCount >= maxEmbeddingVectors {
-			statsSkipped.Add(SkipReasonMaxEmbeddings, int(file.Size))
-			continue
-		}
-
-		if file.Size > maxFileSize {
-			statsSkipped.Add(SkipReasonLarge, int(file.Size))
-			continue
-		}
-
-		if isExcludedFilePath(file.Name, excludePatterns) {
-			statsSkipped.Add(SkipReasonExcluded, int(file.Size))
-			continue
-		}
-
-		contentBytes, err := reader.Read(ctx, file.Name)
-		if err != nil {
-			return embeddings.EmbeddingIndex{}, embeddings.EmbedFilesStats{}, errors.Wrap(err, "error while reading a file")
-		}
-
-		if embeddable, skipReason := isEmbeddableFileContent(contentBytes); !embeddable {
-			statsSkipped.Add(skipReason, len(contentBytes))
-			continue
-		}
-
-		// At this point, we have determined that we want to embed this file.
-
-		for _, chunk := range split.SplitIntoEmbeddableChunks(string(contentBytes), file.Name, splitOptions) {
-			if err := addToBatch(chunk); err != nil {
-				return embeddings.EmbeddingIndex{}, embeddings.EmbedFilesStats{}, err
-			}
-			statsEmbeddedByteCount += len(chunk.Content)
-			statsEmbeddedChunkCount += 1
-		}
-		statsEmbeddedFileCount += 1
-	}
-
-	// Always do a final flush
-	if err := flush(); err != nil {
-		return embeddings.EmbeddingIndex{}, embeddings.EmbedFilesStats{}, err
-	}
-
-	stats := embeddings.EmbedFilesStats{
-		Duration:           time.Since(start),
-		EmbeddedBytes:      statsEmbeddedByteCount,
-		EmbeddedFileCount:  statsEmbeddedFileCount,
-		EmbeddedChunkCount: statsEmbeddedChunkCount,
-		SkippedCounts:      statsSkipped.Counts(),
-		SkippedByteCounts:  statsSkipped.ByteCounts(),
-	}
-
-	return index, stats, nil
+	return entries, nil
 }
 
 type FileReadLister interface {
